+++ src/clients/kmail/KMailClient.ts
@@ -41,7 +41,7 @@ export class KMailClient {
         },
         headers: {
           "kumbi-api-key": "Bearer " + this.config.apiKey,
-          "accept-language": this.config.api.lang,
+          lang: this.config.api.lang,
         },
       });
 
@@ -64,7 +64,7 @@ export class KMailClient {
         },
         headers: {
           "kumbi-api-key": "Bearer " + this.config.apiKey,
-          "accept-language": this.config.api.lang,
+          lang: this.config.api.lang,
         },
       });