+++ src/clients/payment/KPaymentClient.ts
@@ -15,12 +15,6 @@ import APIError from "../../errors/APIError";
 export class KPaymentClient {
   private config: IPaymentParams = {
     apiKey: "",
-    provider: {
-      angolan: {
-        seller: "paypay",
-      },
-      type: "angolan",
-    },
     api: {
       lang: "pt",
       version: "v1",
@@ -34,12 +28,12 @@ export class KPaymentClient {
   async makeStripePayment({ ...data }: IMakeStripePaymentParams) {
     try {
       const response: IMakeStripePaymentResponse = await fetchRequest({
-        url: `${APP_CONFIG.PAYMENT.API_BASE_URL}/${this.config.provider.international.seller}`,
+        url: `${APP_CONFIG.PAYMENT.API_BASE_URL}/stripe`,
         method: "post",
         body: data,
         headers: {
           "kumbi-api-key": `Bearer ${this.config.apiKey}`,
-          "accept-language": this.config.api.lang,
+          lang: this.config.api.lang,
         },
       });
 
@@ -52,12 +46,12 @@ export class KPaymentClient {
   async makeCryptoPayment({ ...data }: IMakeCryptoPaymentParams) {
     try {
       const response: IMakeCryptoPaymentResponse = await fetchRequest({
-        url: `${APP_CONFIG.PAYMENT.API_BASE_URL}/${this.config.provider.international.seller}`,
+        url: `${APP_CONFIG.PAYMENT.API_BASE_URL}/crypto`,
         method: "post",
         body: data,
         headers: {
           "kumbi-api-key": `Bearer ${this.config.apiKey}`,
-          "accept-language": this.config.api.lang,
+          lang: this.config.api.lang,
         },
       });
 
@@ -69,22 +63,7 @@ export class KPaymentClient {
 
   async makeAngolanPayment({ ...data }: IMakeAngolanPaymentParams) {
     try {
-      if (
-        this.config.provider.angolan.seller == "paypay" &&
-        data.method == "multicaixa"
-      ) {
-        let message = {
-          pt: "O método de pagamento via multicaixa express está indisponível para PayPay",
-          en: "The Multicaixa Express payment method is unavailable for PayPay",
-        };
-
-        APIError.ErrorMessage({
-          section: "payment",
-          message: this.config.api.lang == "en" ? message.en : message.pt,
-        });
-      }
-
-      const url = `${APP_CONFIG.PAYMENT.API_BASE_URL}/${this.config.provider.angolan}`;
+      const url = `${APP_CONFIG.PAYMENT.API_BASE_URL}/${data.provider}`;
       const response: IMakeAngolanPaymentResponse = await fetchRequest({
         url: url,
         method: "post",
@@ -93,7 +72,7 @@ export class KPaymentClient {
         },
         headers: {
           "kumbi-api-key": `Bearer ${this.config.apiKey}`,
-          "accept-language": this.config.api.lang,
+          lang: this.config.api.lang,
         },
       });